Add additional `*` -> `+` quantifier mutations
- Adds mutation from `/a*?/` to `/a+?/`
- Adds mutation from `/a*+/` to `/a++/`
