+++ lib/mutant.rb
@@ -93,7 +93,7 @@ require 'mutant/mutator/node/regexp/capture_group'
 require 'mutant/mutator/node/regexp/character_type'
 require 'mutant/mutator/node/regexp/end_of_line_anchor'
 require 'mutant/mutator/node/regexp/end_of_string_or_before_end_of_line_anchor'
-require 'mutant/mutator/node/regexp/greedy_zero_or_more'
+require 'mutant/mutator/node/regexp/zero_or_more'
 require 'mutant/mutator/node/literal'
 require 'mutant/mutator/node/literal/boolean'
 require 'mutant/mutator/node/literal/range'
