--- meta/regexp/regexp_greedy_zero_or_more.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-Mutant::Meta::Example.add :regexp_greedy_zero_or_more do
-  source '/\d*/'
-
-  singleton_mutations
-  regexp_mutations
-
-  mutation '/\d/'
-  mutation '/\d+/'
-  mutation '/\D*/'
-end
